# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/api/genres":{"get":{"tags":["Genres"],"summary":"Get Genres","operationId":"get_genres_api_genres_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Genre"},"type":"array","title":"Response Get Genres Api Genres Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["Genres"],"summary":"Add New Genre","operationId":"add_new_genre_api_genres_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGenre"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Genre"}}}},"200":{"description":"OK"},"202":{"description":"Accepted"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"409":{"description":"Conflict"},"422":{"description":"Validation Error"},"500":{"description":"Internal Server Error"}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/genres/{genre_id}":{"patch":{"tags":["Genres"],"summary":"Update Genre","operationId":"update_genre_api_genres__genre_id__patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"genre_id","in":"path","required":true,"schema":{"type":"integer","title":"Genre Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGenre"}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Genre"}}}},"201":{"description":"Created"},"202":{"description":"Accepted"},"400":{"description":"Bad Request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"},"409":{"description":"Conflict"},"422":{"description":"Validation Error"},"500":{"description":"Internal Server Error"}}}},"/api/auth/login":{"post":{"tags":["Auth"],"summary":"Login","operationId":"login_api_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/auth/refresh":{"post":{"tags":["Auth"],"summary":"Refresh Token","operationId":"refresh_token_api_auth_refresh_post","parameters":[{"name":"refresh_token","in":"query","required":true,"schema":{"type":"string","title":"Refresh Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/users/register":{"post":{"tags":["Users"],"summary":"Register User","operationId":"register_user_api_users_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/users/me":{"get":{"tags":["Users"],"summary":"Get Me","operationId":"get_me_api_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"patch":{"tags":["Users"],"summary":"Update My Data","operationId":"update_my_data_api_users_me_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/users/me/password":{"patch":{"tags":["Users"],"summary":"Update My Password","operationId":"update_my_password_api_users_me_password_patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdatePassword"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/users/{username}":{"get":{"tags":["Users"],"summary":"Get User","operationId":"get_user_api_users__username__get","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserProfile"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/films":{"get":{"tags":["Films"],"summary":"Get Film List","operationId":"get_film_list_api_films_get","responses":{"200":{"description":"Films retrieved successfully","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/FilmSummary"},"type":"array","title":"Response Get Film List Api Films Get"}}}},"404":{"description":"No films found"}}},"post":{"tags":["Films"],"summary":"Add New Film","operationId":"add_new_film_api_films_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_add_new_film_api_films_post"}}},"required":true},"responses":{"201":{"description":"Film created successfully","content":{"application/json":{"schema":{}}}},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/films/{film_id}":{"get":{"tags":["Films"],"summary":"Get Film Details","operationId":"get_film_details_api_films__film_id__get","parameters":[{"name":"film_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Film Id"}}],"responses":{"200":{"description":"Film retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FilmDetail"}}}},"404":{"description":"Film not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/user-films/{film_id}":{"post":{"tags":["User Films"],"summary":"Add User Film","operationId":"add_user_film_api_user_films__film_id__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"film_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Film Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserAddFilm"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserFilmResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["User Films"],"summary":"Update User Film","operationId":"update_user_film_api_user_films__film_id__patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"film_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Film Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserAddFilm"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserFilmResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/user-films/{username}":{"get":{"tags":["User Films"],"summary":"Get User Film List","operationId":"get_user_film_list_api_user_films__username__get","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserFilmResponse"},"title":"Response Get User Film List Api User Films  Username  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reviews/{film_id}":{"post":{"tags":["Reviews"],"summary":"Write Review","operationId":"write_review_api_reviews__film_id__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"film_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Film Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reviews/{review_id}":{"delete":{"tags":["Reviews"],"summary":"Delete Review","operationId":"delete_review_api_reviews__review_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"review_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Review Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Reviews"],"summary":"Update Review","operationId":"update_review_api_reviews__review_id__patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"review_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Review Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewUpdate"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Reviews"],"summary":"Get Review","operationId":"get_review_api_reviews__review_id__get","parameters":[{"name":"film_id","in":"query","required":true,"schema":{"type":"string","format":"uuid","title":"Film Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reviews/{movie_id}":{"get":{"tags":["Reviews"],"summary":"Get Movie Reviews","operationId":"get_movie_reviews_api_reviews__movie_id__get","parameters":[{"name":"movie_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Movie Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Review"},"title":"Response Get Movie Reviews Api Reviews  Movie Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reviews/{username}":{"get":{"tags":["Reviews"],"summary":"Get User Reviews","operationId":"get_user_reviews_api_reviews__username__get","parameters":[{"name":"username","in":"path","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Review"},"title":"Response Get User Reviews Api Reviews  Username  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/reviews/{review_id}/react":{"put":{"tags":["Reviews"],"summary":"React A Review","operationId":"react_a_review_api_reviews__review_id__react_put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"review_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Review Id"}},{"name":"reaction_type","in":"query","required":true,"schema":{"$ref":"#/components/schemas/ReactionType"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Reaction"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Reviews"],"summary":"Delete Reaction","operationId":"delete_reaction_api_reviews__review_id__react_delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"review_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Review Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_add_new_film_api_films_post":{"properties":{"film":{"type":"string","title":"Film"},"images":{"items":{"type":"string","format":"binary"},"type":"array","title":"Images"}},"type":"object","required":["film","images"],"title":"Body_add_new_film_api_films_post"},"Body_login_api_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_api_auth_login_post"},"CreateGenre":{"properties":{"genre_name":{"type":"string","title":"Genre Name"}},"type":"object","required":["genre_name"],"title":"CreateGenre"},"FilmDetail":{"properties":{"title":{"type":"string","title":"Title"},"release_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Release Date"},"air_status":{"type":"string","title":"Air Status"},"film_type":{"type":"string","title":"Film Type"},"episode_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Episode Count"},"rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Rating"},"synopsis":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Synopsis"},"rating_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Rating Count"},"user_rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"User Rating"},"genres":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Genres"},"images":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Images"}},"type":"object","required":["title","air_status","film_type"],"title":"FilmDetail"},"FilmSummary":{"properties":{"title":{"type":"string","title":"Title"},"release_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Release Date"},"air_status":{"type":"string","title":"Air Status"},"film_type":{"type":"string","title":"Film Type"},"episode_count":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Episode Count"},"rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Rating"}},"type":"object","required":["title","air_status","film_type"],"title":"FilmSummary"},"Genre":{"properties":{"id":{"type":"integer","title":"Id"},"genre_name":{"type":"string","maxLength":255,"minLength":1,"title":"Genre Name"}},"type":"object","required":["genre_name"],"title":"Genre"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Reaction":{"properties":{"id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Id"},"user_id":{"type":"string","format":"uuid","title":"User Id"},"review_id":{"type":"string","format":"uuid","title":"Review Id"},"reaction_type":{"$ref":"#/components/schemas/ReactionType"}},"type":"object","required":["user_id","review_id","reaction_type"],"title":"Reaction"},"ReactionType":{"type":"string","enum":["like","dislike"],"title":"ReactionType"},"Review":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"user_film_id":{"type":"integer","title":"User Film Id"},"rating":{"type":"integer","maximum":10.0,"minimum":1.0,"title":"Rating"},"comment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Comment"},"like_count":{"type":"integer","title":"Like Count","default":0},"dislike_count":{"type":"integer","title":"Dislike Count","default":0},"created_at":{"type":"string","format":"date-time","title":"Created At"},"last_updated_at":{"type":"string","format":"date-time","title":"Last Updated At"}},"type":"object","required":["user_film_id","rating"],"title":"Review"},"ReviewCreate":{"properties":{"rating":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Rating"},"comment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Comment"}},"type":"object","title":"ReviewCreate"},"ReviewCreateResponse":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"film_id":{"type":"string","format":"uuid","title":"Film Id"},"rating":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Rating"},"comment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Comment"},"created_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Created At"}},"type":"object","required":["id","film_id"],"title":"ReviewCreateResponse"},"ReviewUpdate":{"properties":{"id":{"type":"string","format":"uuid","title":"Id"},"film_id":{"type":"string","format":"uuid","title":"Film Id"},"rating":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Rating"},"comment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Comment"},"last_updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Last Updated At"}},"type":"object","required":["id","film_id"],"title":"ReviewUpdate"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token"],"title":"Token"},"UserAddFilm":{"properties":{"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"},"progress":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Progress"}},"type":"object","title":"UserAddFilm"},"UserFilmOut":{"properties":{"film_id":{"type":"string","format":"uuid","title":"Film Id"},"status":{"type":"string","title":"Status"},"progress":{"type":"integer","title":"Progress"}},"type":"object","required":["film_id","status","progress"],"title":"UserFilmOut"},"UserFilmResponse":{"properties":{"film_id":{"type":"string","format":"uuid","title":"Film Id"},"film_title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Film Title"},"film_release_date":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Film Release Date"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"},"progress":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Progress"}},"type":"object","required":["film_id"],"title":"UserFilmResponse"},"UserProfile":{"properties":{"username":{"type":"string","title":"Username"},"display_name":{"type":"string","title":"Display Name"},"bio":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bio"},"films":{"anyOf":[{"items":{"$ref":"#/components/schemas/UserFilmOut"},"type":"array"},{"type":"null"}],"title":"Films"}},"type":"object","required":["username","display_name"],"title":"UserProfile"},"UserRegister":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"display_name":{"type":"string","title":"Display Name"}},"type":"object","required":["username","password","display_name"],"title":"UserRegister"},"UserResponse":{"properties":{"username":{"type":"string","title":"Username"},"display_name":{"type":"string","title":"Display Name"},"bio":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bio"},"is_private":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Private"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"last_updated_at":{"type":"string","format":"date-time","title":"Last Updated At"}},"type":"object","required":["username","display_name","bio","is_private","created_at","last_updated_at"],"title":"UserResponse"},"UserUpdate":{"properties":{"display_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Display Name"},"bio":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bio"},"is_private":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Private"}},"type":"object","title":"UserUpdate"},"UserUpdatePassword":{"properties":{"password":{"type":"string","title":"Password"},"new_password":{"type":"string","title":"New Password"}},"type":"object","required":["password","new_password"],"title":"UserUpdatePassword"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/auth/login"}}}}}}